move showToastSuccess, handleApiError -> toast.js
